requirements.txt
# libs/h5py-3.9.0-cp39-cp39-manylinux_2_17_aarch64.manylinux2014_aarch64.whl
# libs/h5py-3.11.0-cp39-cp39-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
# Required dependencies
alembic
flask~=3.0
werkzeug~=3.0
flask-restx
flask-cors
requests~=2.28.1
urllib3~=1.26.12
sqlalchemy
scikit-learn
tensorflow
sagemaker
psycopg2-binary
celery

# Optional dependencies: Can remove if unused
gunicorn
python-dotenv